refactor(indexer): move indexes to separate module
Move the SpIndex struct and related data structures to a new
`indexes.rs` module.

This reorganization improves code structure by separating index-related
functionality from the main library file.

No functional changes were made.
